Conflicts:
	cascader-gui/hosts/level5-s.txt

+++ cascader-gui/lib/cascader-gui/client-gui.rb
@@ -14,7 +14,6 @@ class ClientGui
         host = cascader_re.match(cascader)[2]
         topics = cascader_re.match(cascader)[3].split(':').map { |topic| topic.capitalize }
 
-	### FIX ME        
         lab = HostManager.name_lab(HostManager.lookup_lab(host))
         
         topics.map! do |topic|
@@ -58,6 +57,7 @@ class ClientGui
         lab_hosts = HostManager.get_hosts_lab(@user.floor, @user.lab)
         floor_hosts = HostManager.get_hosts_floor(@user.floor)
         lab_cascaders = []
+        lab_cascaders_map = []
         floor_cascaders = []
         
         lab_hosts.each do |hostname|
@@ -95,23 +95,39 @@ class ClientGui
         end
         
         (floor_cascaders & lab_cascaders).each { |dup| floor_cascaders.delete(dup) }
+	
+	lab_cascaders_map = lab_cascaders.slice(0..-1)
+	
+	unless lab_cascaders.length > 0        
+	        lab_cascaders.map! { |cascader| format_cascader(cascader) }
+		output = "There are #{lab_cascaders.length} cascaders in this lab:\n"
+        	lab_cascaders.each { |cascader| output << cascader+"\n" }
         
-        lab_cascaders.map! { |cascader| format_cascader(cascader) }
-	output = "There are #{lab_cascaders.length} cascaders in this lab:\n"
-        lab_cascaders.each { |cascader| output << cascader+"\n" }
+        	floor_cascaders.map! { |cascader| format_cascader(cascader) }
+        	output <<  "\nThere are #{floor_cascaders.length} other cascaders on this floor:\n"
+        	floor_cascaders.each { |cascader| output << cascader+"\n" }
+	        
+	        dialog = Gtk::MessageDialog.new(
+	       		parent,
+	      		Gtk::Dialog::MODAL,
+	        	Gtk::MessageDialog::INFO,
+			Gtk::MessageDialog::BUTTONS_OK,
+               	message=output
+        	)
+        	dialog.run
+        	dialog.destroy
+	end
         
-        floor_cascaders.map! { |cascader| format_cascader(cascader) }
-        output <<  "\nThere are #{floor_cascaders.length} other cascaders on this floor:\n"
-        floor_cascaders.each { |cascader| output << cascader+"\n" }
-
-        dialog = Gtk::MessageDialog.new(
-	       	parent,
-	      	Gtk::Dialog::MODAL,
-	        Gtk::MessageDialog::INFO,
-		Gtk::MessageDialog::BUTTONS_OK,
-               message=output
-        )
-        dialog.run
-        dialog.destroy
+        if lab_cascaders.length > 0
+        	if @user.lab == "at5n"
+	        	at5n_map(@user.hostname,lab_cascaders_map)
+	        elsif @user.lab == "at5w"
+        		at5w_map(@user.hostname,lab_cascaders_map)
+        	elsif @user.lab == "at5s"
+	        	at5s_map(@user.hostname,lab_cascaders_map)
+	        elsif @user.lab == "at5e"
+	        	at5e_map(@user.hostname,lab_cascaders_map)
+	        end
+	end
     end
 end #class
